Adjust imports and separate Library from App
