+++ src/EventTimeline.css
@@ -1,10 +1,4 @@
-* {
-    font-size: 10px;
-    box-sizing: border-box;
-    padding: 0;
-    margin: 0;
-  }
-  
-  #event-timeline {
-    font-family:'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
-  }
+#event-timeline {
+    font-family: Arial, sans-serif;
+
+}
